+++ Tests/src/Mocks/WebClientMock.h
@@ -62,6 +62,5 @@ public:
             const csp::common::String& MediaType),
         (override));
 
-protected:
     MOCK_METHOD(void, Send, (csp::web::HttpRequest & Request), (override));
 };